# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Dasma Pro - Digital Wedding & Event Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get dashboard statistics for an event
      parameters:
        - name: eventId
          in: query
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Dashboard statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List all events for the current user
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"

    post:
      operationId: createEvent
      tags: [events]
      summary: Create a new event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Created event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "403":
          description: Subscription limit reached
  /events/{id}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get a single event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "404":
          description: Not found

    patch:
      operationId: updateEvent
      tags: [events]
      summary: Update an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventUpdate"
      responses:
        "200":
          description: Updated event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"

    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted

  # ─── GUESTS ───────────────────────────────────────────────────────────────
...
  /checkin:
    post:
      operationId: checkIn
      tags: [checkins]
      summary: Check in a guest via QR token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckInInput"
      responses:
        "200":
          description: Check-in result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckInResult"
        "404":
          description: Token not found
...
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: List all users (admin)
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: number
        - name: limit
          in: query
          required: false
          schema:
            type: number
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserList"
...
  /admin/users/{userId}/subscription:
    patch:
      operationId: adminUpdateUserSubscription
      tags: [admin]
      summary: Update a user's subscription (admin)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriptionUpdate"
      responses:
        "200":
          description: Updated subscription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
components:
  schemas:
...
    Subscription:
      type: object
      properties:
        plan:
          type: string
          enum: [basic, pro, custom]
        eventLimit:
          type: ["number", "null"]
        currentEventCount:
          type: number
        price:
          type: ["number", "null"]
        features:
          type: array
          items:
            type: string
      required:
        - plan
        - currentEventCount
        - features
    SubscriptionUpdate:
      type: object
      properties:
        plan:
          type: string
          enum: [basic, pro, custom]
      required:
        - plan
    DashboardStats:
      type: object
      properties:
        totalGuests:
          type: number
        invitationsSent:
          type: number
        confirmed:
          type: number
        declined:
          type: number
        pending:
          type: number
        checkedIn:
          type: number
        totalTables:
          type: number
        totalCapacity:
          type: number
        seatedGuests:
          type: number
        unseatedConfirmed:
          type: number
      required:
        - totalGuests
        - invitationsSent
        - confirmed
        - declined
        - pending
        - checkedIn
        - totalTables
        - totalCapacity
        - seatedGuests
        - unseatedConfirmed
...
    Event:
      type: object
      properties:
        id:
          type: number
        userId:
          type: string
        name:
          type: string
        date:
          type: string
        time:
          type: ["string", "null"]
        venue:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        dressCode:
          type: ["string", "null"]
        phoneContact:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, active, completed, cancelled]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - userId
        - name
        - date
        - status
        - createdAt
        - updatedAt
    EventInput:
      type: object
      properties:
        name:
          type: string
        date:
          type: string
        time:
          type: string
        venue:
          type: string
        address:
          type: string
        description:
          type: string
        dressCode:
          type: string
        phoneContact:
          type: string
      required:
        - name
        - date
    EventUpdate:
      type: object
      properties:
        name:
          type: string
        date:
          type: string
        time:
          type: string
        venue:
          type: string
        address:
          type: string
        description:
          type: string
        dressCode:
          type: string
        phoneContact:
          type: string
        status:
          type: string
          enum: [draft, active, completed, cancelled]
...
    CheckInInput:
      type: object
      properties:
        qrToken:
          type: string
        eventId:
          type: number
      required:
        - qrToken
        - eventId
    CheckInResult:
      type: object
      properties:
        success:
          type: boolean
        guestName:
          type: string
        tableName:
          type: ["string", "null"]
        seatNumber:
          type: ["number", "null"]
        status:
          type: string
        alreadyCheckedIn:
          type: boolean
      required:
        - success
        - guestName
        - status
        - alreadyCheckedIn
...
    AdminUserList:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/AdminUser"
        total:
          type: number
        page:
          type: number
        limit:
          type: number
      required:
        - users
        - total
        - page
        - limit
    AdminUser:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        role:
          type: string
        subscriptionPlan:
          type: string
        eventCount:
          type: number
        createdAt:
          type: string
      required:
        - id
        - email
        - role
        - subscriptionPlan
        - eventCount
        - createdAt
...